Even More SC2015 Stuff
So I updated the viewer to have a high res, and a low res rendering
method. the low res is currently not working, because I need to get the
positions of each / figure out how to grab the positions properly with
point sprites. it does however work if you draw it instanced with each
vert as a sprite (it looks really cool).