Changed the sphere defaults
I’m about to change a good portion of code and I wanted to set up a
base for reasonable changes, that I can go back to just in case
something screws up. I changed the base number of bodies to 64000, for
simplicity sake and I made the particles a bit closer together. I also
tested out the simulation on a 550ti and I can get about 150^3 (3.375
million) particles on screen at the same time, without it getting too
laggy.