Updated some names
I updated the names of some of the files in here to reflect the recent
decision to remove 3d models (spheres as of now) from the project. I
also updated the readme to reflect the removal of obj2opengl.pl (which
is an AWESOME tool btw). Another thing that was done, was to remove
certain parameters from the VBO/VAO for the spheres, because they were
completely and totally unnecessary, as well as a waste of space for
single-point sprites.