Recording Video
I have set up a very basic video creation workflow by taking pictures
and converting them to video using ffmpeg. STB_Image does not store
images “properly”, but flipped on the Y axis. lines 70 - 76 on the main
takes care of that by flipping the pixel array. I hope to integrate the
ffmpeg library sometime soon as well as get the GUI in here as well,
which is on another repo. I may want to move the screen capture /
recording to the camera class for a cleaner main function.