Winter Cleanup
Yes I have said that I was going to clean up the code for the longest
time, but I never have the time to do it. Between work and school, I
can never catch my breath in order to do this needed task. I’m
currently in the process of getting coverity configured for TravisCI
and that will hopefully allow me to pick out any dead variables /
functions.