Last feature commit
This is the last commit that I will be doing for primary features of
the viewer, anything past this point is maintenance or implementing
JSON / Doing something with the velocities. The JSON implementation
would be for the setup file and would be a quick patch versus a major
change. It will allow us to have different types of simulation run in
the viewer, but I’ll explain more if the time comes. I was not able to
get the sub buffer working as it is an OpenGL 4.4+ feature and I want
to keep all three platforms active (OS X does not go above 4.3 easily
[or at all, I haven’t experimented with NVIDIA drivers yet]). Also
after further investigation, I believe that the framerate drop when
recording is due to file IO on STB’s side (not blaming, I need to
figure out a better way). A possible fix would be to multithread the
application and push the pixel data off to another processor / thread
while I keep rendering in the foreground. With all that being said, I
will be passing off this project to the Tarleton Particle Modeling
Group for continued support. Bye for now.