Update CMakeLists.txt
Minor change to support VisualStudio 2017's new cmake feature, which is amazing. This auto copies out the Viewer-Assets folder to the proper directory. It probably needs another IF statement to say whether or not its debug or release, but I'm lazy and will figure that out later. Don't expect to see major changes to the Viewer, but small tweaks and potential improvements for higher resolution setups / changing resolution while the app is running (preset resolutions in the 16:9 / 16:10 ratio range [I'm not a monster]).