4k resolution
Finally got around to adding in support for different resolutions. you have to manually change it in the code, but the option is still there. Thinking about adding in a little thing to set the resolution before GL inits. Travis CI build will fail because packages are old.  need to fix that soon.