Regenerated GLAD
Went ahead and regenerated the GLAD config and created a 'static' version of it.
This will update GL to the latest standard and MAY fix the windows build.
The windows build borked whenever I made some changes to how the resolution was handled.
For some reason the windows build broke while the linux version stayed the same, go figure.